third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/textfieldselection/selection-start-end-extra-expected.txt

This is a testharness.js-based test.
PASS Setting defaultValue in a textarea should move the cursor to the end
PASS Setting defaultValue in a textarea with a value should NOT make any difference
PASS Setting textContent in a textarea should move selection{Start,End} to the end
PASS Adding children to a textarea should move selection{Start,End} to the end
PASS Removing children from a textarea should update selection{Start,End}
FAIL Setting the same value (with different newlines) in a textarea should NOT update selection{Start,End} assert_equals: expected 3 but got 14
PASS Setting value to a shorter string than defaultValue should correct the cursor position
PASS Shortening value by turning the input type into 'url' should correct selection{Start,End}
FAIL Shortening value by turning the input type into 'color' and back to 'text' should correct selection{Start,End} assert_equals: expected 7 but got 9
PASS Resetting a value to a shorter string than defaultValue should correct the cursor position
Harness: the test ran to completion.
